+++ src/main.js
@@ -5,6 +5,9 @@ const util = require("node:util");
 const execFilePromisified = util.promisify(
   require("node:child_process").execFile
 );
+const createMenuTemplate = require("./menu");
+
+const isDev = !app.isPackaged;
 
 const FileExtensions = {
   flextext: ".flextext",
@@ -33,106 +36,7 @@ const createWindow = () => {
   mainWindow.loadURL(MAIN_WINDOW_WEBPACK_ENTRY);
 
   // menubar
-  const menuTemplate = [
-    {
-      label: 'File',
-      submenu: [
-        {
-          label: 'Exit',
-          role: 'quit', // Closes the application
-        },
-      ],
-    },
-    {
-      label: 'View',
-      submenu: [
-        {
-            label: 'Zoom In',
-            accelerator: 'CmdOrCtrl+Plus',
-            click: () => {
-                const focusedWindow = BrowserWindow.getFocusedWindow();
-                if (focusedWindow) {
-                    focusedWindow.webContents.setZoomLevel(
-                        focusedWindow.webContents.getZoomLevel() + 0.5
-                    );
-                }
-            }
-        },
-        {
-            label: 'Zoom Out',
-            accelerator: 'CmdOrCtrl+-',
-            click: () => {
-                const focusedWindow = BrowserWindow.getFocusedWindow();
-                if (focusedWindow) {
-                    focusedWindow.webContents.setZoomLevel(
-                        focusedWindow.webContents.getZoomLevel() - 0.5
-                    );
-                }
-            }
-        },
-        {
-            label: 'Reset Zoom',
-            accelerator: 'CmdOrCtrl+0',
-            click: () => {
-                const focusedWindow = BrowserWindow.getFocusedWindow();
-                if (focusedWindow) {
-                    focusedWindow.webContents.setZoomLevel(0);
-                }
-            }
-        }
-      ]
-    },
-    {
-      label: 'Help',
-      submenu: [
-        {
-          label: 'Getting Started',
-          click: () => {
-            dialog.showMessageBox({
-              type: 'info',
-              title: 'Getting Started',
-              message: 'Getting Started',
-              detail: `1. Select the file(s) to be converted.\n` +
-                      `2. Choose the desired input and output file formats.\n` +
-                      `3. Configure mappings or additional settings, if applicable.\n` +
-                      `4. Click "Convert" and specify the destination for the converted file.`,
-              buttons: ['OK'],
-            });
-          },
-        },
-        {
-          label: 'Rebabel Documentation',
-          click: () => {
-            shell.openExternal('https://github.com/mr-martian/rebabel-format/tree/master/docs');
-          },
-        },
-        { type: 'separator' },
-        {
-          label: 'About',
-          click: () => {
-            dialog.showMessageBox({
-              type: 'info',
-              title: 'About Gap App',
-              message: 'Gap App',
-              detail: `Version: ${app.getVersion()}\n` +
-              `\n` +
-              `Welcome to Gap App! This tool is designed to simplify file format conversions with just a few clicks. Follow the steps below to quickly convert your files and customize the output to your needs.\n` +
-              `\n` +
-              `Linguists rely on a variety of specialized software tools to document and preserve endangered languages. A common challenge they face is the need to transfer language data between different applications, a process that currently lacks an automated and user-friendly solution. Gap App provides an efficient and intuitive solution for converting NLP output files into language data formats compatible with software such as Fieldworks Language Explorer (FLEx) and ELAN.\n` +
-              `\n` +
-              `Team Members:\n` +
-              `- Joseph Barron: Backend Developer/Scrum Master\n` +
-              `- Adassa Coimin: Frontend/Backend Developer\n` +
-              `- Matthew Denslinger: Frontend Developer\n` +
-              `- Elizabeth Thorner: Backend Developer/Project Manager\n` +
-              `- Darren Wang: Frontend Developer`,
-              buttons: ['OK'],
-            });
-          },
-        },
-      ],
-    },
-  ];
+  const menuTemplate = createMenuTemplate(isDev);
 
   // Build the menu from the template
   const menu = Menu.buildFromTemplate(menuTemplate);
